requirements.txt
Flask==2.3.2
boto3==1.26.0
Flask-Cors==3.0.10
python-dotenv==1.0.0
pandas==2.0.0
requests==2.28.1
